+++ src/app/project/[projectId]/develop-recipe/page.tsx
@@ -2,7 +2,7 @@ import { DevelopRecipe } from '@/components/templates/develop-recipe/DevelopReci
 import { getProject } from '@/service/get-project-service';
 import { notFound, redirect } from 'next/navigation';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/lib/next-auth/auth-options';
 
 const DevelopRecipePage = async ({
   params,
